Move Vorbis and Opus decoding out of WebMReader
