+++ dom/media/webm/moz.build
@@ -13,6 +13,7 @@ EXPORTS += [
 ]
 
 UNIFIED_SOURCES += [
+    'AudioDecoder.cpp',
     'SoftwareWebMVideoDecoder.cpp',
     'WebMBufferedParser.cpp',
     'WebMDecoder.cpp',
